requirements.txt
Flask==2.1.1
Flask_Cors==3.0.10
pycryptodome==3.14.1
gunicorn